requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
openpyxl>=3.1.0
google-search-results>=2.4.0
python-dotenv>=1.0.0
validators>=0.20.0
PyPDF2>=3.0.0
chardet>=5.0.0
fake-useragent>=1.4.0
google-generativeai>=0.3.0
googlesearch-python>=1.2.3
selenium>=4.15.0
